chore(pr) Moved code out of dispatch.tsx.
